Создание простой задачи для контроля их выполнения, инициализация очереди. Корректировка тестов, для совместного использования генератора уникальных ID.

+++ Multithreading/src/main/java/blockingqueue/UniqueIdGenerator.java
@@ -14,7 +14,7 @@ public class UniqueIdGenerator
 {
 	private final static AtomicInteger generatorValue = new AtomicInteger(0);
 
-	public static long nextId()
+	public static int nextId()
 	{
 		return generatorValue.incrementAndGet();
 	}
